requirements.txt
configobj~=5.0.6
requests~=2.28.1
psycopg2-binary~=2.9.3
pytz~=2020.5
--index-url https://${GIT_TOKEN_USER}:${GIT_TOKEN}@gitlab.com/api/v4/projects/32772862/packages/pypi/simple
fins

